requirements.txt
fastapi
uvicorn
mangum
tensorflow==2.18.0
pillow
python-multipart
fastapi.middleware.cors